fix(core): adapt to move of locations in ansys.dpf.core.common
